classes_homework_java/src/main/java/Calculator.java
public class Calculator {

    private int firstNumber;

    private int secondNumber;

    public Calculator (int firstNumber, int secondNumber){
        this.firstNumber = firstNumber;
        this.secondNumber = secondNumber;
    }

    public int add(){
        return firstNumber + secondNumber;
    }

    public int subtract(){
        return firstNumber - secondNumber;
    }

    public int multiply(){
        return firstNumber * secondNumber;
    }

    public double divide(){
        return (double) firstNumber /secondNumber;
    }
}
